Create DAY 13 : CASE STATEMENTS
Today, the focus was on using case statements in Verilog, which are very similar to the switch statements found in languages like C but with some key differences. They are useful in combinational circuits when you need to choose between multiple possible values based on a single input, making them an efficient alternative to chained if-else statements, especially when there are numerous conditions.